Truncate ld_desc
Due to SL not being able to handle the expansion of the ld_desc column, current workaround is to truncate the ld_desc string before insert. Combined with earlier addition that will pause put tag function to prevent Teamwork servers refusing to update tags, this should reduce our chance of importing a duplicate to near zero.

Will revisit and explore inserting a notes entry per line-item in future to properly solve the problem